GG-19062 CacheGroupsMetricsRebalanceTest.testRebalanceEstimateFinishTime fixed.

+++ modules/core/src/main/java/org/apache/ignite/internal/processors/cache/CacheMetricsImpl.java
@@ -1121,7 +1121,7 @@ public class CacheMetricsImpl implements CacheMetrics {
         long rate = rebalancingKeysRate.getRate();
 
         return rate <= 0 ? -1L :
-            ((getKeysToRebalanceLeft() / rate) * REBALANCE_RATE_INTERVAL) + U.currentTimeMillis();
+            (long)(((double)getKeysToRebalanceLeft() / rate) * REBALANCE_RATE_INTERVAL) + U.currentTimeMillis();
     }
 
     /** {@inheritDoc} */
